Rewrite "ephemeral swap" rc.d script
The old script was designed to handle EC2 Xen instances (with disks
appearing as /dev/xbd*), and also broke in FreeBSD 11 and later due
to the SWAPMETA zone disappearing (as a result of changes in the
FreeBSD kernel swap code).

The new script works with Nitro instances and "nvme" disks, making
use of the /dev/aws/disk/ephemeral tree to get a list of ephemeral
disks and using /var/db/ec2_ephemeral_diskseen and the disk serial
numbers to ensure that we only allocate swap space on "new" disks.

Because the old script was nonfunctional for so long, the new script
has a different name and different enabling variable
(ec2_ephemeral_swap_enable rather than ec2_ephemeralswap_enable) in
order to ensure that people upgrading to the latest code aren't
surprised by suddenly having it turned on if they had a "legacy"
setting in /etc/rc.conf.